+++ wallets/imtoken-extension/src/imToken.ts
@@ -1,0 +1,5 @@
+import { IMTokenWallet, imTokenWalletInfo } from './extension';
+
+const imTokenWallet = new IMTokenWallet(imTokenWalletInfo);
+
+export const wallets = [imTokenWallet];