+++ wallets/leap-capsule-social-login/src/client.ts
@@ -1,8 +1,18 @@
 /* eslint @typescript-eslint/no-explicit-any: 0 */ // --> OFF
 
-import { StdSignDoc } from '@cosmjs/amino';
-import { SignOptions, SignType } from '@cosmos-kit/core';
-import { DirectSignDoc, WalletClient } from '@cosmos-kit/core';
+import {
+  encodeEd25519Pubkey,
+  encodeSecp256k1Pubkey,
+  pubkeyType,
+  StdSignature,
+  StdSignDoc,
+} from '@cosmjs/amino';
+import {
+  DirectSignDoc,
+  SignOptions,
+  SignType,
+  WalletClient,
+} from '@cosmos-kit/core';
 import { CapsuleProvider } from '@leapwallet/cosmos-social-login-capsule-provider';
 import { SignDoc } from 'cosmjs-types/cosmos/tx/v1beta1/tx';
 
@@ -58,6 +68,44 @@ export class CosmosCapsuleClient implements WalletClient {
     signOptions?: SignOptions
   ) {
     return this.loginProvider.signAmino(chainId, signer, signDoc, signOptions);
+  }
+
+  async signArbitrary(
+    chainId: string,
+    signer: string,
+    data: string | Uint8Array
+  ): Promise<StdSignature> {
+    const account = await this.getAccount(chainId);
+    if (!account) {
+      throw new Error(`Wallet not connected to chain ${chainId}`);
+    }
+    const pubkey = (() => {
+      switch (account.algo) {
+        case 'secp256k1':
+          return encodeSecp256k1Pubkey(account.pubkey);
+        case 'ed25519':
+          return encodeEd25519Pubkey(account.pubkey);
+        default:
+          throw new Error('sr25519 public key algorithm is not supported');
+      }
+    })();
+
+    const signature = await this.loginProvider.signArbitrary(
+      chainId,
+      signer,
+      data
+    );
+
+    return {
+      signature,
+      pub_key: {
+        type:
+          account.algo === 'secp256k1'
+            ? pubkeyType.secp256k1
+            : pubkeyType.ed25519,
+        value: pubkey.value,
+      },
+    };
   }
 
   async signDirect(