+++ packages/example/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.27.2](https://github.com/cosmology-tech/cosmos-kit/compare/example@2.27.1...example@2.27.2) (2024-10-14)
+
+**Note:** Version bump only for package example
+
+
+
+
+
+## [2.27.1](https://github.com/cosmology-tech/cosmos-kit/compare/example@2.27.0...example@2.27.1) (2024-10-09)
+
+**Note:** Version bump only for package example
+
+
+
+
+
+# [2.27.0](https://github.com/cosmology-tech/cosmos-kit/compare/example@2.26.0...example@2.27.0) (2024-10-09)
+
+**Note:** Version bump only for package example
+
+
+
+
+
+# [2.26.0](https://github.com/cosmology-tech/cosmos-kit/compare/example@2.25.1...example@2.26.0) (2024-10-09)
+
+**Note:** Version bump only for package example
+
+
+
+
+
+## [2.25.1](https://github.com/cosmology-tech/cosmos-kit/compare/example@2.25.0...example@2.25.1) (2024-09-30)
+
+**Note:** Version bump only for package example
+
+
+
+
+
 # [2.25.0](https://github.com/cosmology-tech/cosmos-kit/compare/example@2.24.0...example@2.25.0) (2024-08-01)
 
 **Note:** Version bump only for package example