+++ packages/react-lite/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.15.1](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/react-lite@2.15.0...@cosmos-kit/react-lite@2.15.1) (2024-10-14)
+
+**Note:** Version bump only for package @cosmos-kit/react-lite
+
+
+
+
+
+# [2.15.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/react-lite@2.14.0...@cosmos-kit/react-lite@2.15.0) (2024-10-09)
+
+**Note:** Version bump only for package @cosmos-kit/react-lite
+
+
+
+
+
+# [2.14.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/react-lite@2.13.0...@cosmos-kit/react-lite@2.14.0) (2024-10-09)
+
+**Note:** Version bump only for package @cosmos-kit/react-lite
+
+
+
+
+
 # [2.13.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/react-lite@2.12.2...@cosmos-kit/react-lite@2.13.0) (2024-07-11)
 
 **Note:** Version bump only for package @cosmos-kit/react-lite