+++ packages/walletconnect/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.12.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/walletconnect@2.11.0...@cosmos-kit/walletconnect@2.12.0) (2024-10-09)
+
+**Note:** Version bump only for package @cosmos-kit/walletconnect
+
+
+
+
+
+# [2.11.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/walletconnect@2.10.1...@cosmos-kit/walletconnect@2.11.0) (2024-10-09)
+
+**Note:** Version bump only for package @cosmos-kit/walletconnect
+
+
+
+
+
 ## [2.10.1](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/walletconnect@2.10.0...@cosmos-kit/walletconnect@2.10.1) (2024-06-28)
 
 **Note:** Version bump only for package @cosmos-kit/walletconnect