+++ wallets/leap-capsule-social-login/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.1](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/leap-capsule-social-login@0.12.0...@cosmos-kit/leap-capsule-social-login@0.12.1) (2024-10-09)
+
+**Note:** Version bump only for package @cosmos-kit/leap-capsule-social-login
+
+
+
+
+
+# [0.12.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/leap-capsule-social-login@0.11.0...@cosmos-kit/leap-capsule-social-login@0.12.0) (2024-10-09)
+
+**Note:** Version bump only for package @cosmos-kit/leap-capsule-social-login
+
+
+
+
+
+# [0.11.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/leap-capsule-social-login@0.10.3...@cosmos-kit/leap-capsule-social-login@0.11.0) (2024-10-09)
+
+**Note:** Version bump only for package @cosmos-kit/leap-capsule-social-login
+
+
+
+
+
+## [0.10.3](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/leap-capsule-social-login@0.10.2...@cosmos-kit/leap-capsule-social-login@0.10.3) (2024-09-30)
+
+**Note:** Version bump only for package @cosmos-kit/leap-capsule-social-login
+
+
+
+
+
 ## [0.10.2](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/leap-capsule-social-login@0.10.1...@cosmos-kit/leap-capsule-social-login@0.10.2) (2024-06-28)
 
 **Note:** Version bump only for package @cosmos-kit/leap-capsule-social-login