+++ wallets/trust/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.13.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/trust@2.12.0...@cosmos-kit/trust@2.13.0) (2024-10-09)
+
+**Note:** Version bump only for package @cosmos-kit/trust
+
+
+
+
+
+# [2.12.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/trust@2.11.2...@cosmos-kit/trust@2.12.0) (2024-10-09)
+
+**Note:** Version bump only for package @cosmos-kit/trust
+
+
+
+
+
 ## [2.11.2](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/trust@2.11.1...@cosmos-kit/trust@2.11.2) (2024-06-28)
 
 **Note:** Version bump only for package @cosmos-kit/trust